# OpenWA/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/contacts": {
      "get": {
        "operationId": "ContactController_findAll",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max contacts to return (1–1000, default 1000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of contacts to skip (for paging)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of contacts"
          },
          "400": {
            "description": "Session not ready"
          },
          "404": {
            "description": "Session not found"
          }
        },
        "summary": "Get all contacts for a session",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/sessions/{sessionId}/contacts/profile-pictures": {
      "get": {
        "description": "One request for a whole chat sidebar — avoids the burst of parallel single fetches that would exhaust the per-IP throttle. Engine lookups run 3 at a time; per-id failures return null.",
        "operationId": "ContactController_getProfilePictures",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "Comma-separated contact ids (max 50 used)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Batch-resolve profile picture URLs for up to 50 contacts",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/sessions/{sessionId}/contacts/{contactId}": {
      "get": {
        "operationId": "ContactController_findOne",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID (e.g., 628xxx@c.us)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact details"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "summary": "Get a specific contact by ID",
        "tags": [
          "contacts"
        ]
      },
      "put": {
        "operationId": "ContactController_upsertContact",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID (e.g., 628xxx@c.us)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact saved"
          },
          "400": {
            "description": "Session not active or invalid request"
          }
        },
        "summary": "Save a contact to the account's addressbook, or edit an existing entry",
        "tags": [
          "contacts"
        ]
      },
      "delete": {
        "operationId": "ContactController_deleteContact",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID (e.g., 628xxx@c.us)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted"
          },
          "400": {
            "description": "Session not active"
          }
        },
        "summary": "Remove a contact from the account's addressbook",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/sessions/{sessionId}/contacts/check/{number}": {
      "get": {
        "description": "Returns whether the number is a registered WhatsApp account and its canonical id. Use this to pre-validate a recipient before sending: the send endpoints return 201 on accepting a message even for numbers that are not on WhatsApp, so this is the only way to confirm a new number is reachable before you send to it.",
        "operationId": "ContactController_checkNumber",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "number",
            "required": true,
            "in": "path",
            "description": "Phone number to check (e.g., 628123456789)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Number existence check result"
          }
        },
        "summary": "Check if a phone number exists on WhatsApp",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/sessions/{sessionId}/contacts/{contactId}/profile-picture": {
      "get": {
        "operationId": "ContactController_getProfilePicture",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID (e.g., 628xxx@c.us)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile picture URL"
          }
        },
        "summary": "Get profile picture URL for a contact",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/sessions/{sessionId}/contacts/{contactId}/phone": {
      "get": {
        "operationId": "ContactController_resolvePhone",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID / JID to resolve (e.g., an @lid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resolved phone number (MSISDN digits), or null when the engine cannot map it"
          }
        },
        "summary": "Resolve a contact id (e.g. an @lid) to a phone number — best-effort",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/sessions/{sessionId}/contacts/{contactId}/block": {
      "post": {
        "operationId": "ContactController_blockContact",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID (e.g., 628xxx@c.us)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact blocked"
          }
        },
        "summary": "Block a contact",
        "tags": [
          "contacts"
        ]
      },
      "delete": {
        "operationId": "ContactController_unblockContact",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contactId",
            "required": true,
            "in": "path",
            "description": "Contact ID (e.g., 628xxx@c.us)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact unblocked"
          }
        },
        "summary": "Unblock a contact",
        "tags": [
          "contacts"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenWA API",
    "description": "Open Source WhatsApp API Gateway - Free, Self-Hosted HTTP API",
    "version": "0.13.0",
    "contact": {
      "name": "OpenWA",
      "url": "https://github.com/rmyndharis/OpenWA",
      "email": "yudhi@rmyndharis.com"
    }
  },
...
  "servers": [
    {
      "url": "http://{host}:{port}",
      "description": "OpenWA instance",
      "variables": {
        "host": {
          "default": "localhost"
        },
        "port": {
          "default": "2785",
          "description": "PORT env var"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "X-API-Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "metrics-bearer": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http",
        "description": "METRICS_TOKEN for the GET /api/metrics scrape endpoint"
      }
    },
    "schemas": {
...
      "UpsertContactDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "The contact's first name.",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "description": "The contact's last name. Omit for a single-name contact — WhatsApp allows those.",
            "maxLength": 100
          }
        },
        "required": [
          "firstName"
        ]
      },
...
    }
  },
  "security": [
    {
      "X-API-Key": []
    }
  ]
}